+++ command/command.ts
@@ -1441,7 +1441,7 @@ export class Command<
             });
           }
 
-          if (arg) {
+          if (typeof arg !== "undefined") {
             params.push(arg);
           }
         }