+++ command/test/command/arguments_test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, assertThrowsAsync } from "../../../dev_deps.ts";
+import { assertEquals, assertRejects } from "../../../dev_deps.ts";
 import type { ITypeInfo } from "../../../flags/types.ts";
 import { Command } from "../../command.ts";
 
@@ -28,8 +28,13 @@ Deno.test("valid command argument types", async () => {
   assertEquals(args, ["abc", 123, true, "red"]);
 });
 
+Deno.test("command - arguments - with integer value", async () => {
+  const { args } = await cmd().parse(["abc", "0"]);
+  assertEquals(args, ["abc", 0]);
+});
+
 Deno.test("invalid number command argument type", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd().parse(["abc", "xyz", "true", "red"]);
     },
@@ -39,7 +44,7 @@ Deno.test("invalid number command argument type", async () => {
 });
 
 Deno.test("missing command arguments", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd().parse();
     },
@@ -49,7 +54,7 @@ Deno.test("missing command arguments", async () => {
 });
 
 Deno.test("invalid boolean command argument type", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd().parse(["abc", "123", "xyz", "red"]);
     },
@@ -59,7 +64,7 @@ Deno.test("invalid boolean command argument type", async () => {
 });
 
 Deno.test("invalid custom command argument type", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd().parse(["abc", "123", "true", "xyz"]);
     },