+++ command/test/command/stop_early_test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, assertThrowsAsync } from "../../../dev_deps.ts";
+import { assertEquals, assertRejects } from "../../../dev_deps.ts";
 import { Command } from "../../command.ts";
 
 Deno.test("command stopEarly disable", async () => {
@@ -65,7 +65,7 @@ Deno.test("command stopEarly unknown option", async () => {
     .option("-S, --script-arg2 [value:boolean]", "description ...")
     .action(() => {});
 
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd.parse([
         "-f",