+++ command/test/command/sub_command_test.ts
@@ -1,7 +1,7 @@
 import {
   assertEquals,
+  assertRejects,
   assertThrows,
-  assertThrowsAsync,
 } from "../../../dev_deps.ts";
 import { Command } from "../../command.ts";
 
@@ -86,7 +86,7 @@ Deno.test("command - sub command - nested child command with arguments", async (
 });
 
 Deno.test("command - sub command - sub-command with missing argument", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await command().parse(["sub-command", "input-path"]);
     },
@@ -96,7 +96,7 @@ Deno.test("command - sub command - sub-command with missing argument", async () 
 });
 
 Deno.test("command - sub command - sub-command 2 with missing argument", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await command().parse(["sub-command2", "input-path"]);
     },
@@ -106,7 +106,7 @@ Deno.test("command - sub command - sub-command 2 with missing argument", async (
 });
 
 Deno.test("command - sub command - nested sub-command with missing argument", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await command().parse(["sub-command2", "sub-command3", "input-path"]);
     },
@@ -116,7 +116,7 @@ Deno.test("command - sub command - nested sub-command with missing argument", as
 });
 
 Deno.test("command - sub command - command with empty name", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await new Command()
         .command("")
@@ -135,7 +135,7 @@ Deno.test("command - sub command - override child command", async () => {
 });
 
 Deno.test("command - sub command - duplicate command name", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await new Command()
         .command("foo")