+++ command/test/type/integer_test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, assertThrowsAsync } from "../../../dev_deps.ts";
+import { assertEquals, assertRejects } from "../../../dev_deps.ts";
 import { Command } from "../../command.ts";
 
 const cmd = new Command()
@@ -22,7 +22,7 @@ Deno.test("command - type - integer - with valid value", async () => {
 });
 
 Deno.test("command - type - integer - with argument", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd.parse(["-f", "123", "unknown"]);
     },
@@ -32,7 +32,7 @@ Deno.test("command - type - integer - with argument", async () => {
 });
 
 Deno.test("command - type - integer - with missing value", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd.parse(["-F"]);
     },
@@ -42,7 +42,7 @@ Deno.test("command - type - integer - with missing value", async () => {
 });
 
 Deno.test("command - type - integer - with invalid string value", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd.parse(["-f", "abc"]);
     },
@@ -52,7 +52,7 @@ Deno.test("command - type - integer - with invalid string value", async () => {
 });
 
 Deno.test("command - type - integer - with invalid float value", async () => {
-  await assertThrowsAsync(
+  await assertRejects(
     async () => {
       await cmd.parse(["-f", "1.23"]);
     },