+++ table/utils.ts
@@ -9,30 +9,21 @@ import { stripColor } from "./deps.ts";
 
 export function consumeWords(length: number, content: string): string {
   let consumed = "";
-  const words: string[] = content.split(/ /g);
+  const words: string[] = content.split("\n")[0]?.split(/ /g);
 
   for (let i = 0; i < words.length; i++) {
-    let word: string = words[i];
-    const hasLineBreak = word.indexOf("\n") !== -1;
-
-    if (hasLineBreak) {
-      word = word.split("\n").shift() as string;
-    }
+    const word: string = words[i];
 
     // consume minimum one word
     if (consumed) {
-      const nextLength = stripColor(word).length;
-      const consumedLength = stripColor(consumed).length;
+      const nextLength = strLength(word);
+      const consumedLength = strLength(consumed);
       if (consumedLength + nextLength >= length) {
         break;
       }
     }
 
     consumed += (i > 0 ? " " : "") + word;
-
-    if (hasLineBreak) {
-      break;
-    }
   }
 
   return consumed;
@@ -46,20 +37,37 @@ export function longest(
   rows: ICell[][],
   maxWidth?: number,
 ): number {
-  return Math.max(
-    ...rows.map((row) =>
-      (
-        row[index] instanceof Cell && (row[index] as Cell).getColSpan() > 1
-          ? ""
-          : (row[index]?.toString() || "")
-      )
-        .split("\n")
-        .map((r: string) => {
-          const str = typeof maxWidth === "undefined"
-            ? r
-            : consumeWords(maxWidth, r);
-          return stripColor(str).length || 0;
-        })
-    ).flat(),
-  );
+  const cellLengths = rows.map((row) => {
+    const cell = row[index];
+    const cellValue = cell instanceof Cell && cell.getColSpan() > 1
+      ? ""
+      : cell?.toString() || "";
+
+    return cellValue
+      .split("\n")
+      .map((line: string) => {
+        const str = typeof maxWidth === "undefined"
+          ? line
+          : consumeWords(maxWidth, line);
+
+        return strLength(str) || 0;
+      });
+  }).flat();
+
+  return Math.max(...cellLengths);
 }
+
+export const strLength = (str: string): number => {
+  str = stripColor(str);
+  let length = 0;
+  for (let i = 0; i < str.length; i++) {
+    const charCode = str.charCodeAt(i);
+    // Check for chinese characters: \u4e00 - \u9fa5
+    if (charCode >= 19968 && charCode <= 40869) {
+      length += 2;
+    } else {
+      length += 1;
+    }
+  }
+  return length;
+};